% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regal_align.R
\name{align_embeddings}
\alias{align_embeddings}
\title{Node Representation Alignment}
\usage{
align_embeddings(emb1, emb2)
}
\arguments{
\item{emb1}{A matrix of node representations for the first graph, where each row corresponds to a node.}

\item{emb2}{A matrix of node representations for the second graph, where each row corresponds to a node.}
}
\value{
A vector indicating the aligned node indices from the second graph for each node in the first graph.
}
\description{
This function aligns node representations across two graphs using a k-d tree.
}
