% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/bada.R
\name{bada}
\alias{bada}
\alias{bada.multidesign}
\title{Barycentric Discriminant Analysis}
\usage{
bada(data, y, subject, preproc, ncomp)

\method{bada}{multidesign}(
  data,
  y,
  subject,
  preproc = center(),
  ncomp = 2,
  resdim = 20,
  rescomp = 0,
  ...
)
}
\arguments{
\item{data}{the object (e.g. \code{multidesign}) containing the X and Y data.}

\item{y}{the categorical response variable}

\item{subject}{the subject variable}

\item{preproc}{pre-processing function, defaults to \code{center}}

\item{ncomp}{number of components to estimate}

\item{resdim}{pca dimensionality for residual analysis (only relevant if \code{rescomp} > 0)}

\item{rescomp}{number of final residual components (default = 0, no residual aanalysis)}

\item{...}{arguments to pass through}
}
\description{
A component technique that maximizes the between group variance over a set of variables.
}
\details{
The \code{subject} argument is used to model multilevel structure. When \code{rescomp} > 0, then the group barycenters
will be subtracted from each subject's data and a residual pca-lda will be computed to estimate components of variation
that deviate from the group barycenters. This can be used to capture individual level variation that can be useful
for modeling idiosyncratic structure in the multilevel data.
}
\examples{

X <- matrix(rnorm(50*100), 50, 100)
Y <- tibble(condition=rep(letters[1:5], 10), subject=rep(1:5, each=10))

des <- multivarious::multidesign(X,Y)
bres1 <- bada(des, condition, subject=subject, rescomp=0)
bres2 <- bada(des, condition, subject=subject, resdim=20, rescomp=5)
}
\references{
Abdi, H., Williams, L. J., & Bera, M. (2017). Barycentric discriminant analysis. \emph{Encyclopedia of Social Network Analysis and Mining}, 1-20.
}
