% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bada.R
\name{bootstrap.bada}
\alias{bootstrap.bada}
\title{Bootstrap Resampling for bada Multivariate Models}
\usage{
\method{bootstrap}{bada}(x, data, nboot = 500, alpha = 0.05, ...)
}
\arguments{
\item{x}{A fitted bada model object that has been fit to a training dataset.}

\item{data}{The dataset on which the bootstrap resampling should be performed.}

\item{nboot}{An integer specifying the number of bootstrap resamples to perform (default is 500).}

\item{...}{Additional arguments to be passed to the specific model implementation of \code{bootstrap}.}
}
\value{
A list containing the summarized bootstrap resampled components and scores for the model.
The returned list contains eight elements:
\itemize{
\item \code{boot_scores_mean}: A matrix which is the mean of all bootstrapped scores matrices.
\item \code{boot_scores_sd}: A matrix which is the standard deviation of all bootstrapped scores matrices.
\item \code{boot_scores_upper}: A matrix which is the upper alpha percentile of all bootstrapped scores matrices.
\item \code{boot_scores_lower}: A matrix which is the lower alpha percentile of all bootstrapped scores matrices.
\item \code{boot_lds_mean}: A matrix which is the mean of all bootstrapped loadings matrices.
\item \code{boot_lds_sd}: A matrix which is the standard deviation of all bootstrapped loadings matrices.
\item \code{boot_lds_upper}: A matrix which is the upper alpha percentile of all bootstrapped loadings matrices.
\item \code{boot_lds_lower}: A matrix which is the lower alpha percentile of all bootstrapped loadings matrices.
The dimensions of each matrix in the list correspond to the dimensions of the respective matrices in the input data.
}
}
\description{
Perform bootstrap resampling on a multivariate bada model to estimate the variability
of components and scores.
}
