% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled_diagonalization.R
\name{compute_eigenvectors}
\alias{compute_eigenvectors}
\title{Compute Eigenvectors and Eigenvalues}
\usage{
compute_eigenvectors(L, k_prime)
}
\arguments{
\item{L}{The Laplacian matrix.}

\item{k_prime}{The number of eigenvectors and eigenvalues to compute.}
}
\value{
A list containing the eigenvectors (U_bar) and eigenvalues (Lambda_bar).
}
\description{
This function computes the first k' eigenvectors and eigenvalues of a given Laplacian matrix.
}
