% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{compute_grad_Jp}
\alias{compute_grad_Jp}
\title{Compute the gradient of the orthogonality regularizer Jp}
\usage{
compute_grad_Jp(W)
}
\arguments{
\item{W}{Matrix of weights (d x m)}

\item{m}{Number of reduced dimensions}
}
\value{
Gradient of Jp with respect to W (d x m)
}
\description{
Compute the gradient of the orthogonality regularizer Jp
}
