% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{compute_P}
\alias{compute_P}
\title{Compute the similarity matrix P}
\usage{
compute_P(Y, sigma_P)
}
\arguments{
\item{sigma_P}{Scaling parameter for the Gaussian kernel}

\item{W}{Matrix of weights (d x m)}

\item{X}{Matrix of input data (n x d)}
}
\value{
Matrix P of similarity scores (n x n)
}
\description{
Compute the similarity matrix P
}
