% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled_diagonalization.R
\name{construct_laplacian}
\alias{construct_laplacian}
\title{Construct Laplacian Matrix}
\usage{
construct_laplacian(X, sigma, knn = 5)
}
\arguments{
\item{X}{A numeric matrix representing the data samples.}

\item{sigma}{The bandwidth parameter for the Gaussian kernel.}

\item{knn}{The number of nearest neighbors to consider.}
}
\value{
The normalized Laplacian matrix.
}
\description{
This function constructs the normalized Laplacian matrix from a given data matrix.
}
