% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{covstatis}
\alias{covstatis}
\title{Statis for covariance matrices}
\usage{
covstatis(data, ncomp, normalize, dcenter)
}
\arguments{
\item{data}{a set of covariance matrices}

\item{normalize}{whether to normalize so that the sums-of-squares of each matrix is 1.}

\item{dcenter}{whether to double center the covariance matrix.}

\item{the}{number of componentss to compute}
}
\description{
Statis for covariance matrices
}
