% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasp.R
\name{GRASP_align}
\alias{GRASP_align}
\title{Align two graphs using the GRASP algorithm}
\usage{
GRASP_align(
  G1,
  G2,
  k = 20,
  q = 100,
  t = seq(0.1, 50, length.out = q),
  mu = 0.132
)
}
\arguments{
\item{G1}{An igraph object representing the first graph.}

\item{G2}{An igraph object representing the second graph.}

\item{k}{The number of eigenvectors to use for alignment (default: 20).}

\item{q}{The number of corresponding functions to use (default: 100).}

\item{t}{A vector of time steps for the heat kernel (default: seq(0.1, 50, length.out = q)).}

\item{mu}{The regularization parameter for the optimization problem (default: 0.132).}
}
\value{
A list containing the following elements:
\itemize{
\item alignment: A vector of length |V1| indicating the aligned node indices in G2 for each node in G1.
\item P: The permutation matrix representing the node alignment.
\item M: The alignment matrix for the eigenvectors.
\item C: The diagonal mapping matrix for the corresponding functions.
}
}
\description{
Align two graphs using the GRASP algorithm
}
