% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasp.R
\name{linear_assignment}
\alias{linear_assignment}
\title{Perform linear assignment using the Jonker-Volgenant algorithm}
\usage{
linear_assignment(C1, C2)
}
\arguments{
\item{C1}{The first matrix of size (n x k) for assignment.}

\item{C2}{The second matrix of size (n x k) for assignment.}
}
\value{
A vector of length n indicating the assigned indices.
}
\description{
Perform linear assignment using the Jonker-Volgenant algorithm
}
