% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{linear_sim_embed}
\alias{linear_sim_embed}
\title{Linear Similarity Embedding using ADAM optimization}
\usage{
linear_sim_embed(
  X,
  T,
  M,
  sigma_P = 1,
  ncomp = 2,
  ap = 0.2,
  maxit = 100,
  tol = 1e-08,
  batch_size = 0.1,
  use_cpp = TRUE
)
}
\arguments{
\item{X}{Input data matrix (n x d)}

\item{T}{Target similarity matrix (n x n)}

\item{M}{Mask matrix (n x n)}

\item{sigma_P}{Scaling parameter for the Gaussian kernel}

\item{ncomp}{Number of reduced dimensions}

\item{ap}{Weighting parameter for the orthogonality regularizer}

\item{maxit}{Maximum number of iterations for optimization}

\item{tol}{Tolerance for optimization}
}
\value{
Optimized weight matrix W (d x m)
}
\description{
Linear Similarity Embedding using ADAM optimization
}
