% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa.multiblock}
\alias{mfa.multiblock}
\title{Multiple factor analysis}
\usage{
\method{mfa}{multiblock}(
  data,
  preproc = center(),
  ncomp = 2,
  normalization = c("MFA", "RV", "None", "Frob", "custom"),
  M = NULL,
  A = NULL,
  ...
)
}
\arguments{
\item{data}{a \code{multiblock} object}

\item{preproc}{a preprocessing pipeline, default is \code{center()}}

\item{ncomp}{the number of components to estimate}

\item{normalization}{the normalization method: MFA, RV, RV-MFA, or None (see details).}

\item{M}{when \code{normalization} is \code{custom}, a weight (sparse or dense) matrix for the rows}

\item{A}{when \code{normalization} is \code{custom}, a weight (sparse or dense) matrix for the columns}
}
\description{
principal component analysis for multiple blocks of data collected over th same instances
}
\examples{

X <- replicate(5, { matrix(rnorm(10*10), 10, 10) }, simplify=FALSE)
res <- mfa(X, ncomp=3, normalization="MFA")
p <- project_block(res, X[[1]], 1)
stopifnot(ncol(scores(res)) == 3)

labs <- letters[1:10]
cfier <- classifier(res, new_data=do.call(cbind, X), labels=labs)
pred <- predict(cfier, X[1:2,])
cfier2 <- classifier(res, new_data=X[[2]], labels=labs, colind=res$block_indices[[2]])
pred2 <- predict(cfier2, X[1:2,res$block_indices[[2]]])
}
\references{
Abdi, H., Williams, L. J., & Valentin, D. (2013). Multiple factor analysis: principal component analysis for multitable and multiblock data sets.
\emph{Wiley Interdisciplinary reviews: computational statistics}, 5(2), 149-179.
}
