% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{mfa}
\alias{mfa}
\title{Multiple Factor Analysis}
\usage{
mfa(data, preproc, ncomp, normalization, A, M)
}
\arguments{
\item{data}{the object (e.g. \code{multidesign}) containing the X and Y data.}

\item{preproc}{pre-processing function, defaults to \code{center}}

\item{ncomp}{number of components to estimate}

\item{normalization}{the normalization method: MFA, RV, RV-MFA, or None (see details).}

\item{A}{custom weight matrix for the columns}

\item{M}{custom weight matrix for the rows}
}
\description{
principal component analysis for multiple blocks of data collected over th same instances
}
\references{
Abdi, H., Williams, L. J., & Valentin, D. (2013). Multiple factor analysis: principal component analysis for multitable and multiblock data sets.
\emph{Wiley Interdisciplinary reviews: computational statistics}, 5(2), 149-179.
}
