% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasp.R
\name{normalized_laplacian}
\alias{normalized_laplacian}
\title{Compute the normalized Laplacian matrix of a graph}
\usage{
normalized_laplacian(G)
}
\arguments{
\item{G}{An igraph object representing the graph.}
}
\value{
The normalized Laplacian matrix of the graph.
}
\description{
Compute the normalized Laplacian matrix of a graph
}
