% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled_diagonalization.R
\name{optimization}
\alias{optimization}
\title{Coupled Diagonalization Optimization}
\usage{
optimization(U_bars, Lambda_bars, F_mats, G_mats, k, k_prime, mu_c, mu_d)
}
\arguments{
\item{U_bars}{A list of matrices containing the eigenvectors of the Laplacian matrices.}

\item{Lambda_bars}{A list of diagonal matrices containing the eigenvalues of the Laplacian matrices.}

\item{F_mats}{A list of matrices representing the coupling constraints.}

\item{G_mats}{A list of matrices representing the decoupling constraints.}

\item{k}{The desired reduced dimension.}

\item{mu_c}{The coupling strength parameter.}

\item{mu_d}{The decoupling strength parameter.}
}
\value{
A list of optimal transformation matrices A_i.
}
\description{
This function performs the optimization step of the coupled diagonalization method.
}
