% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{optimize_W}
\alias{optimize_W}
\title{Main function to perform optimization using ADAM}
\usage{
optimize_W(X, T, M, initial_W, sigma_P = 1, ap = 0.2, maxit = 200, tol = 1e-08)
}
\arguments{
\item{X}{Input data matrix (n x d)}

\item{T}{Target similarity matrix (n x n)}

\item{M}{Mask matrix (n x n)}

\item{initial_W}{Initial weight matrix (d x m)}

\item{sigma_P}{Scaling parameter for the Gaussian kernel}

\item{ap}{Weighting parameter for the orthogonality regularizer}

\item{maxit}{Maximum number of iterations for optimization}

\item{tol}{Tolerance for optimization}
}
\value{
Optimized weight matrix W (d x m)
}
\description{
Main function to perform optimization using ADAM
}
