% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipca.R
\name{pcpc_test}
\alias{pcpc_test}
\title{Title}
\usage{
pcpc_test(Cs, pc_candi, m = 100)
}
\arguments{
\item{m}{the number of replicates for parametric simulation.}

\item{gamma}{Candidates of common eigenvectors (\hat{\Gamma}_{candi})}

\item{cov_data}{a list of covariance matrices}

\item{ns}{the number of samples to estimate each covariance matrix.}
}
\value{
A list of three objects. Object "cpc_index" tells which columns of gamma are CPC estimates. Object "summary_mat"
gives the information about the sequential testing procedure. Object "cpc" gives the CPC estimates.
}
\description{
Title
}
\examples{
See the example in pcpc_est.R.
}
