% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bada.R
\name{project.bada}
\alias{project.bada}
\title{new sample projection}
\usage{
\method{project}{bada}(x, new_data, block)
}
\arguments{
\item{x}{The model fit, typically an object of class bi_projector or any other class that implements a project method}

\item{new_data}{A matrix or vector of new observations with the same number of columns as the original data. Rows represent observations and columns represent variables}

\item{block}{the \code{character} id for the subject/block to project on.
This is required for block-specific pre-processing.
If missing, the group average pre-processing parameters are used.}
}
\description{
Project one or more samples of onto a subspace.
}
