% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bada.R
\name{project.bada}
\alias{project.bada}
\title{project}
\usage{
project.bada(x, new_data, subject)
}
\arguments{
\item{subject}{the \code{character} id for the subject to project on.
This is required for subject-specific pre-processing.}
}
\description{
project
}
