% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{project_cov}
\alias{project_cov}
\title{project covariance matrix}
\usage{
project_cov(x, new_data, ...)
}
\arguments{
\item{x}{the model fit}

\item{new_data}{the covariance matrix}

\item{...}{extra args}
}
\description{
project a new covariance matrix onto a component fit
}
