% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regal_align.R
\name{regal_align}
\alias{regal_align}
\title{REGAL: REpresentation-based Graph ALignment}
\usage{
regal_align(
  A1,
  A2,
  k = 2,
  p = 100,
  d = 128,
  attributes1 = NULL,
  attributes2 = NULL
)
}
\arguments{
\item{A1}{An adjacency matrix representing the first graph.}

\item{A2}{An adjacency matrix representing the second graph.}

\item{k}{The maximum number of hops to consider for node identity extraction.}

\item{p}{The number of landmark nodes to use in xNetMF.}

\item{d}{The dimensionality of the learned node representations.}

\item{attributes1}{A matrix of node attributes for the first graph (optional).}

\item{attributes2}{A matrix of node attributes for the second graph (optional).}
}
\value{
A list with two elements:
\itemize{
\item embeddings: A matrix of aligned node representations, where each row corresponds to a node in G1 or G2.
\item alignment: A vector indicating the aligned node indices from G2 for each node in G1.
}
}
\description{
This function implements the REGAL algorithm for aligning nodes across two graphs
based on their learned representations. It consists of three main steps:
\enumerate{
\item Node identity extraction
\item Efficient similarity-based representation learning using xNetMF
\item Fast node representation alignment using a k-d tree
}
}
