% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regal_align.R
\name{xNetMF}
\alias{xNetMF}
\title{xNetMF: Cross-Network Matrix Factorization}
\usage{
xNetMF(X, p, d)
}
\arguments{
\item{X}{A matrix of node identity features, where each row corresponds to a node.}

\item{p}{The number of landmark nodes to use.}

\item{d}{The dimensionality of the learned node representations.}
}
\value{
A matrix of learned node representations, where each row corresponds to a node.
}
\description{
This function learns node representations using the xNetMF method, which efficiently
factorizes an approximated similarity matrix based on the Nyström method.
}
