% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bada.R
\name{bada}
\alias{bada}
\title{Barycentric Discriminant Analysis}
\usage{
bada(
  y,
  subject,
  data,
  ncomp = 2,
  preproc = center(),
  resdim = 20,
  rescomp = 2,
  ...
)
}
\arguments{
\item{subject}{the subject formula}

\item{data}{the \code{multidesign} containing the X and Y data.}

\item{ncomp}{number of components to estimate}

\item{preproc}{pre-processing function, defaults to \code{center}}

\item{...}{arguments to pass through}

\item{formula}{the fixed effects formula}

\item{A}{the column constraints}

\item{M}{the row constraints}
}
\description{
A component technique that maximizes the between group variance over a set of variables.
}
\details{
The \code{subject} argument can be used to model multilevel structure.
}
\examples{

X <- matrix(rnorm(50*100), 50, 100)
Y <- tibble(condition=rep(letters[1:5], 10), subject=rep(1:5, each=10))

des <- multivarious::multidesign(X,Y)
bada(condition, subject=subject, data=des)
}
\references{
Abdi, H., Williams, L. J., & Bera, M. (2017). Barycentric discriminant analysis. \emph{Encyclopedia of Social Network Analysis and Mining}, 1-20.
}
