% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bada.R
\name{project.bada}
\alias{project.bada}
\title{new sample projection}
\usage{
\method{project}{bada}(x, new_data, block)
}
\arguments{
\item{x}{the model fit}

\item{new_data}{a matrix or vector of new observations(s) with the same number of columns as the original data.}

\item{block}{the \code{character} id for the subject/block to project on.
This is required for block-specific pre-processing.
If missing, the group average pre-processing parameters are used.}
}
\description{
Project one or more samples of onto a subspace.
}
