% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trace_ratio}
\alias{trace_ratio}
\title{trace_ratio optimization}
\usage{
trace_ratio(A, B, ncomp = 2, eps = 1e-06, maxiter = 100)
}
\arguments{
\item{A}{the numerator matrix}

\item{B}{the denominator matrix}
}
\description{
trace_ratio optimization
}
\examples{
data(iris)
X <- as.matrix(iris[,1:4])
Y <- iris[,5]
A <- between_class_scatter(X, Y, colMeans(X))
B <- within_class_scatter(X, Y)
ret <- trace_ratio(A,B, ncomp=3)
}
\keyword{internal}
